\name{mantel.correlog-package}
\alias{mantel.correlog-package}
\docType{package}
\title{ Mantel correlogram }
\description{
Function \code{\link{mantel.correlog}} computes a multivariate Mantel correlogram.
}
\details{
\tabular{ll}{
Package: \tab mantel.correlog\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2009-08-23\cr
License: \tab Copyleft\cr
LazyLoad: \tab yes\cr
}
Function \code{\link{mantel.correlog}} computes a multivariate Mantel correlogram.
}

\author{ Pierre Legendre }

\references{
Legendre, P. & L. Legendre. 1998. Numerical ecology, 2nd English edition. Elsevier Science BV, Amsterdam.

Mantel, N. 1967. The detection of disease clustering and a generalized regression approach. Cancer Res. 27: 209-220.

Oden, N. L. & R. R. Sokal. 1986. Directional autocorrelation: an extension of spatial correlograms to two dimensions. Syst. Zool. 35: 608-617.

Sokal, R. R. 1986. Spatial data analysis and historical processes. 29-43 in: E. Diday et al. [eds.] Data analysis and informatics, IV. North-Holland, Amsterdam.
}

\keyword{ package }
